🧪 test: merge test modules
- removed constants module
- removed more unecessary initialization functions
- moved remaining constants to the main test module
