+++ src/db/tests/mod.rs
@@ -5,19 +5,36 @@ use std::time::Duration;
 
 use tokio::time::{Instant, sleep_until};
 
-mod constants;
 mod item_params;
 mod join_params;
 mod receipts_params;
 mod receipts_users_params;
 mod user_params;
 
-use constants::*;
-
 #[sqlx::test]
 async fn test_db_conn() {
     let conn = get_db().await;
     assert!(conn.is_ok());
+}
+
+pub const TEST_USERS: [&str; 3] = ["Thing", "Noodle", "Jon"];
+
+pub const TEST_ITEMS: [(&str, f64, i64); 3] = [
+    ("PB Pretzel", 4.99, 2),
+    ("Slamin' Salmon", 9.49, 1),
+    ("Chips and Dip", 5.55, 3),
+];
+
+pub fn intermediate_totals() -> Vec<HashMap<i64, Decimal>> {
+    vec![
+        HashMap::from([(3, dec!(9.98))]),
+        HashMap::from([(2, dec!(9.49))]),
+        HashMap::from([(2, dec!(8.32)), (3, dec!(8.32))]),
+    ]
+}
+
+pub fn expected_totals() -> HashMap<i64, Decimal> {
+    HashMap::from([(3, dec!(18.30)), (2, dec!(17.81))])
 }
 
 async fn init_users(conn: &SqlitePool) -> Result<Vec<StoreUser>> {
